Several changes to make way for the Physically Based Shading changes.
- Changed instances to be drawn using UBOs instead of per-instance
  attributes (HUGE perf gain!)
- Added KFrustum for frustum culling. Now when rendering the instance
  manager will cull based on the current camera's view-frustum.
- Added a DebugGBufferPass which allows quick switching to see different
  parts of the GBuffer. For now it must be turned off/on whether it is
  compiled into the code.
- Currently using defines to state which BRDF is in use. However this is
  very inefficient, and should be switched to shader subroutines. The only
  reason this wasn't the initial action is because GLES 3.1 doesn't
  support shader subroutines. So I will have to special-case it. On PC I
  will allow subroutines and dynamic BRDF switching, but on ES I will have
  to lock-in on a particular BRDF. Compiling all of the combinations of
  shaders for each BRDF just isn't realistic. It takes forever on PC, I
  can only imagine the time it would take on an Embedded System.
- Materials have been reworked to maintain their own UBO. They now support
  Diffuse, Specular, Roughness, and Fresnel. Considering support for other
  material properties.
- Instances also have been reworked to use UBOs, they have their own for
  the three common transformation matrices.
- InstanceManager sorts based on Mesh, then based on Material. A UBO
  binding is only done if absolutely needed.
- GLSL no longer relies on Karma-Specific bindings. Instead there is a
  Bindings.glsl that all uniforms can reference for their initial binding
  location. C++ code also references this header to keep things in-sync.
- Several different equations have been implemented for microfacet BRDFs,
  however a clear winner cannot be stated until some more advanced
  functionality is present.

+++ Karma/ktrianglepartition.h
@@ -17,7 +17,7 @@ public:
   KTrianglePartition(KPointCloud const &cloud);
 protected:
   KPointCloud const &m_pointCloud;
-  KVector3D const &point(int idx) const;
+  KVector3D const &point(size_t idx) const;
 };
 
 inline KTrianglePartition::KTrianglePartition(const KPointCloud &cloud) :
@@ -26,7 +26,7 @@ inline KTrianglePartition::KTrianglePartition(const KPointCloud &cloud) :
   // Intentionally Empty
 }
 
-inline const KVector3D &KTrianglePartition::point(int idx) const
+inline const KVector3D &KTrianglePartition::point(size_t idx) const
 {
   return m_pointCloud[idx - 1];
 }
